+++ packages/webawesome/docs/docs/resources/changelog.md
@@ -33,7 +33,7 @@ Web Awesome follows <a href="https://semver.org/" class="appearance-plain">Seman
 
 :::added
 
-- Added the `<wa-checkbox-group>` component for grouping related checkboxes or switches under a shared label and hint.
+- Added the `<wa-checkbox-group>` component
 - Added `leaf-multiple` as a new `selection` option for `<wa-tree>`, allowing multiple leaf nodes to be selected while parent nodes only expand and collapse.
 
 :::